requirements.txt
streamlit
pandas
numpy
scikit-learn
plotly
pydeck
matplotlib
joblib
Pillow
streamlit-player